pyproject.toml
[project]
name = "training-program-001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly-express>=0.4.1",
    "streamlit>=1.47.1",
]
